moved sidebar opener to new file
